User profiles are now persisted to disk using MongoDB.
The code was refactored to accomodate the new chanages. The `User` model is used as the primary
interface to the database. Currently the passwords are stored without hashing.

The payload of JWT is now just the User identifier within the database. The life time of the
token was extended to 7 days. We may want to create a collection called `session`. This would
allow us to revoke tokens, say when the user logs out.

The `POST /users` and `POST /sessions` endpoints can be accessed without authorization. Otherwise,
the client can neither create a session nor an user.