Passwords are now hashed using the BCrypt algorithm.
The salt is automatically generated by BCrypt.js after 10 rounds. If we change this in the future,
should we rehash all the passwords in the database? I am not sure about this.

We are currently passing callbacks to all the asynchronous calls. I am starting to see the first
signs of the infamous "callback hell". We should start making use of the Promise API and friends.