Refactored the authentication controller.
The `jwtCheck` and `requireRole` middleware were moved out of the authentication controller.
They were refactored to be more generic. The authentication controller now uses the `unless`
middleware, which was also moved to the `/middleware` directory, to conditionally prevent
`jwtCheck` and `requireRole` middleware from executing based on the request.

The Joi package was added as a dependency in this commit. We will be using it to validate
incoming request body.