Created the POST /accounts endpoint.
The controller modules now export the `attachRoutes` function which accepts a router. The endpoints
implemented by the controllers are attached to this router. Thus, router configurations can be done
in a single location.

The parameters received by the POST /accounts endpoint are verified using Joi. The user name for an
account must be unique within the customers organization. The `ownerId` attribute introduced to
the `Account` model helps us ensure this constraint. The endpoint responds with 401 (Bad Request)
if the specified user name already exists. Otherwise, it inserts the new account in the database
and responds with 201 (Created).