Added the GET /accounts endpoint.
The GET /accounts endpoint accepts a filter. The filter is sanitized using a schema via Joi. All
the parameters have not been defined yet. Only pagination related parameters are supported.
Pagination is implemented using a customized version of `mongoose-pagination-v2`. The unit tests
were updated to use Jest.

The endpoint responds with 401 (Bad Request) if the specified filter is invalid. Otherwise, it
returns the accounts owned by the current user and responds with 201 (Created). The maximum
number of accounts that can be fetched per request is limited to 100. We need to implement other
security measures such as limiting the number of page requests a certain IP address can make.
When we implement API keys, we will be vulnerable to data breach otherwise.

MongoDB presets were added as development dependency. The Jest configuration was updated
appropriately. Unit and integration tests were implemented as well. However, they are not complete.
Further, the tests fail if Jest is run on watch mode. This could be related to the in-memory
MongDB instance we spawn for testing.

`app.js` was updated to export the server instance from the module. The `start.js` module is
now responsible for listening to incoming requests. This refactor was done to accomodate the
integration tests using SuperTest.