Added the GET /accounts/:identifier endpoint.
The GET /accounts/:identifier endpoint returns the details of the account with the specified
identifier. An account created by one user should be hidden from another user. Right now,
I have not written tests for this constraint. Further, the endpoint responds with 404 (Not Found)
if the specified identifier is invalid. Otherwise, it returns the account owned by the current
user and responds with 200 (OK).

A few updates were made to the test cases. Jest is now configured to run tests serially.
This prevents multiple test cases messing up the in-memory MongoDB instance.