Added the PUT /accounts/:identifier endpoint.
The PUT /accounts/:identifier endpoint updates and returns the account with the specified
identifier. The parameters received by the endpoint are verified using Joi. An account
created by one user should be hidden from another user. Right now, I have not written tests
for this constraint. Further, the endpoint responds with 404 (Not Found) if the specified
identifier is invalid. Otherwise, it returns the account owned by the current user and HTTP
status code 200 (OK).