Added the DELETE /accounts/:identifier endpoint.
The DELETE /accounts/:identifier endpoint marks the account with the specified
identifier as deleted. The account is technically not deleted from the database.
An account created by one user cannot be deleted by another user. Right now, I
have not written tests for this endpoint. Further, the endpoint responds with 404
(Not Found) if the specified identifier is invalid. Otherwise it returns the an empty
response with HTTP status code 204 (No Content).

All the endpoints in the accounts module were updated to acknowledge deleted accounts.
Thus, clients cannot access deleted accounts.