Created the POST /plans endpoint.
The parameters received by the POST /plans endpoint are validated using Joi. The plan name must
be unique within the customers organization. The `ownerId` attribute in the `Plan` model helps
us ensure this constraint. The endpoint responds with 401 (Bad Request) if the specified plan name
already exists. Otherwise the server persists the new plan to the database and responds with
201 (Created).