Added the GET /plans/:identifier endpoint.
The GET /plans/:identifier endpoint returns the details of the plan with the specified
identifier. A plan created by one user should be hidden from another user. Right now,
I have not written tests for this constraint. Further, the endpoint responds with 404 (Not Found)
if the specified identifier is invalid. Otherwise it returns the plan owned by the current
user with the HTTP status code 200 (OK).