Added the PUT /plans/:identifier endpoint.
The PUT /plans/:identifier endpoint updates and returns the plan with the specified
identifier. The parameters received by the endpoint are verified using Joi. A plan
created by one user should be hidden from another user. Right now, I have not written tests
for this constraint. Further, the endpoint responds with 404 (Not Found) if the specified
identifier is invalid. Otherwise, it returns the plan owned by the current user and HTTP
status code 200 (OK).