Added the DELETE /plans/:identifier endpoint.
The DELETE /plans/:identifier endpoint marks the plan with the specified
identifier as deleted. The plan is technically not deleted from the database.
A plan created by one user cannot be deleted by another user. Right now, I
have not written tests for this endpoint. Further, the endpoint responds with 404
(Not Found) if the specified identifier is invalid. Otherwise it returns the an empty
response with HTTP status code 204 (No Content).

All the endpoints in the plans module were written to acknowledge deleted plans.
Thus, clients cannot access deleted plans.