Added the GET /subscriptions/:id endpoint.
The GET /subscriptions/:id endpoint returns the details of the subscription with the specified
identifier. A subscription created by one user should be hidden from another user. Right now,
I have not written tests for this endpoint. Further, the endpoint responds with 404 (Not Found)
if the specified identifier is invalid. Otherwise it returns the subscription owned by the current
user with the HTTP status code 200 (OK).