Added the GET /subscriptions endpoint.
The GET /subscriptions endpoint accepts a filter. The filter is sanitized using a schema via Joi. All
the parameters have not been defined yet. Only pagination related parameters are supported.

The endpoint responds with 401 (Bad Request) if the specified filter is invalid. Otherwise it
returns the subscriptions owned by the current user and the HTTP status code 201 (Created). The maximum
number of subscriptions that can be fetched per request is limited to 100. We need to implement other
security measures such as limiting the number of page requests a certain IP address can make.
When we implement API keys, we will be vulnerable to data breach otherwise.