Added the GET /invoices/:id endpoint.
The GET /invoices/:id endpoint returns the details of the invoice with the specified
identifier. An invoice created by one user should be hidden from another user. Right now,
I have not written tests for this endpoint. Further, the endpoint responds with 404 (Not Found)
if the specified identifier is invalid. Otherwise it returns the invoice owned by the current
user with the HTTP status code 200 (OK).