Added the GET /transactions/:id endpoint.
The GET /transactions/:id endpoint returns the details of the transaction with the specified
identifier. A transaction created by one user should be hidden from another user. Right now,
I have not written tests for this endpoint. Further, the endpoint responds with 404 (Not Found)
if the specified identifier is invalid. Otherwise it returns the transaction owned by the current
user with the HTTP status code 200 (OK).