Implemented an utility function for generating invoice numbers.
We need a scalable way to generate invoice numbers. For example, if we have 100 server instances
running, how do we maintain a sequential order for each company? Do we hit the database whenever
we want an invoice? That would be bad because of the latency and the load on the database for such
a simple task.

We also considered using block identifiers, where each server instance gets (say three
alphanumeric unique identifier). Now we have a counter that starts at 1 for each instance. Assume,
an instance with identifier ABC. Now the counter is at 1. So the invoice would be ABC-0001. The next
invoice number generated by that instance would be ABC-0002. Similarly, if there was another instance
running, it would generate ABD-0001, ABD-0002, and so on. The problem is, this identifier has to be
unique for each company, too. To make things even worse, there is no guarantee that the instances
are invoked serially.  Therefore, we decided to go with Redis, which would be just like the database
approach, except much faster.

We are using the Node Redis package to connect to the Redis server. The utility function accepts the
prefix and the identifier of the company and returns a promise. Redis is used to store an integer for
a given company identifier. Something like an hash map.

Redis has the concept of expiring. We take advantage of this to reset the integer at the end of the day.
We ask Redis to expire they when the `incr` returns 1. We use the formula `total seconds in 24 hours -
seconds passed since 12 AM` to evaluate the TTL.

For two companies, Redis would look something like:
```
5f2a596584beac1179c3dca0 -> 1
5f32d133e69f652728345ef4 -> 5
```