The GET /accounts endpoint now accepts an optional search text parameter.
The specified search text parameter is used to filter accounts based on user name, first name, and
last name. The index for the same must be created in the database.

It should be noted that there is a bug in this commit. Interestingly, we can ship the product with
the bug without causing any harm. Basically, the search does not work when the text includes both
the first and the last names. For example, if there is an account called "Anushka Madyanam", you
can search for "anushka" or "madyanam". However, if you searched for "anushka madyanam" you would
not get anything. This is because we consider the search text as a single regex pattern.